Fix golangci-lint errors: shadow, contextcheck, unconvert, rowserrcheck
- Remove unnecessary int64 conversion in diskspace_unix.go
- Pass context to bootstrap goroutine for contextcheck compliance
- Rename shadowing err variables throughout codebase to avoid govet shadow warnings
- Add missing rows.Err() check in state.go GetPendingChunks

+++ internal/cache/diskspace_unix.go
@@ -11,5 +11,5 @@ func (c *Cache) getDiskFreeSpace() (int64, error) {
 		return 0, err
 	}
 	// Available blocks * block size
-	return int64(stat.Bavail) * int64(stat.Bsize), nil
+	return int64(stat.Bavail) * stat.Bsize, nil
 }
